modif

+++ simpan-pengajuan.php
@@ -9,17 +9,29 @@
 	$tanggal		= $_POST['tanggal'];
 
 
-	$query		= "INSERT INTO tbl_pengajuan (id_pengajuan,id_customer,id_user,id_barang,status,lama_angsuran,jml_angs,tgl_pengajuan) VALUES ('','$customer','1','$barang','Baru','$angsuran','$angsuran','$tanggal')";
+	$cek = "SELECT * FROM tbl_pengajuan WHERE id_customer = '$customer'";
+	
+	$qr = mysql_query($cek) or die(mysql_error());
 
-	$input=mysql_query($query) or die (mysql_error());
+	$exist = mysql_num_rows($qr);
+	// echo $exist;
 
-if ($input){
-	echo "Berhasil menambahkan data";
-	echo '<META HTTP-EQUIV="REFRESH" CONTENT = "1; URL=data-pengajuan.php">';
-}   
+	if($exist == 0){
+		$query		= "INSERT INTO tbl_pengajuan (id_pengajuan,id_customer,id_user,id_barang,status,lama_angsuran,jml_angs,tgl_pengajuan) VALUES ('','$customer','1','$barang','Baru','$lama','$angsuran','$tanggal')";
 
+		$input=mysql_query($query) or die (mysql_error());
 
+		if ($input){
+			echo "Berhasil menambahkan data";
+			echo '<META HTTP-EQUIV="REFRESH" CONTENT = "1; URL=data-pengajuan.php">';
+		}   
 
+	}else{
 
-
+		echo "<script> 
+				alert('Maaf anda tidak bisa melakukan pengajuan')
+			</script>";
+		echo '<META HTTP-EQUIV="REFRESH" CONTENT = "1; URL=data-pengajuan.php">';
+		   	
+	}	
 ?>
